# public/api-docs/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Фудмаркет API",
		"version": "1.0.0",
		"description": "REST API интернет-магазина «Фудмаркет» (food-market22.ru). Авторизация — cookie-сессия better-auth."
	},
	"servers": [
		{
			"url": "/",
			"description": "Текущий хост"
		},
		{
			"url": "https://food-market22.ru",
			"description": "Продакшен"
		}
	],
	"components": {
		"securitySchemes": {
			"cookieAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "better-auth.session_token",
				"description": "Cookie-сессия better-auth. Устанавливается после логина."
			}
		},
...
	},
...
	"security": [],
	"paths": {
...
		"/api/auth/login": {
			"post": {
				"tags": [
					"Auth"
				],
				"summary": "Вход по телефону и паролю (кастомная сессия)",
				"description": "Проверяет пароль (bcrypt), создаёт сессию и ставит cookie `session`.",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": [
									"phoneNumber",
									"password"
								],
								"properties": {
									"phoneNumber": {
										"type": "string"
									},
									"password": {
										"type": "string",
										"format": "password"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Успешный вход, ставит cookie session"
					},
					"401": {
						"description": "Неверный пароль"
					},
					"404": {
						"description": "Пользователь не найден"
					},
					"500": {
						"description": "Ошибка сервера"
					}
				}
			}
		},
...
		"/api/auth/reset-phone-pass": {
			"post": {
				"tags": [
					"Auth"
				],
				"summary": "Сбросить пароль по номеру телефона",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": [
									"phoneNumber",
									"newPassword"
								],
								"properties": {
									"phoneNumber": {
										"type": "string"
									},
									"newPassword": {
										"type": "string",
										"format": "password"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Пароль обновлён"
					},
					"400": {
						"description": "Не хватает данных"
					},
					"404": {
						"description": "Пользователь не найден"
					},
					"500": {
						"description": "Ошибка сервера"
					}
				}
			}
		},
...
	}
}